+++ src/js/src/views/ScorchRuns.vue
@@ -33,7 +33,7 @@
           <vue-terminal :wsPath="terminal.loc"></vue-terminal>
         </section>
         <footer class="modal-card-foot buttons is-right">
-          <div v-if="terminal.ro || !roleAllowed('scorch', 'post')">
+          <div v-if="terminal.ro || !scorchTerminalWriteAllowed()">
             <b-tooltip
               label="this will close but not exit the terminal"
               type="is-light is-left"
@@ -85,7 +85,10 @@
   import axiosInstance from '@/utils/axios.js';
   import { useErrorNotification } from '@/utils/errorNotif';
   import { usePhenixStore } from '@/store.js';
-  import { roleAllowed, scorchControlAllowed } from '@/utils/rbac.js';
+  import {
+    scorchControlAllowed,
+    scorchTerminalWriteAllowed,
+  } from '@/utils/rbac.js';
 
   import ScorchKey from '@/components/scorch/ScorchKey.vue';
   import ScorchRun from '@/components/scorch/ScorchRun.vue';
@@ -93,7 +96,7 @@
 
   export default {
     setup() {
-      return { roleAllowed, scorchControlAllowed };
+      return { scorchControlAllowed, scorchTerminalWriteAllowed };
     },
 
     components: {